Refactoring: GeoDataRiverSurvey::DeleteRiverPathPointCommand is moved to separated files.
