+++ libs/geodata/riversurvey/riversurvey.pro
@@ -146,6 +146,7 @@ HEADERS += gd_riversurvey_global.h \
            gridcreatingconditionriversurveyinterface.h \
            private/geodatariversurvey_addextensioncommand.h \
            private/geodatariversurvey_changeselectioncommand.h \
+           private/geodatariversurvey_deleteriverpathpointcommand.h \
            private/geodatariversurvey_setdisplaysettingcommand.h \
            private/geodatariversurveycrosssectionwindow_datatabledelegate.h \
            private/geodatariversurveycrosssectionwindow_impl.h \
@@ -193,6 +194,7 @@ SOURCES += geodatarivercrosssection.cpp \
            geodatariversurveyvtkexporter.cpp \
            private/geodatariversurvey_addextensioncommand.cpp \
            private/geodatariversurvey_changeselectioncommand.cpp \
+           private/geodatariversurvey_deleteriverpathpointcommand.cpp \
            private/geodatariversurvey_setdisplaysettingcommand.cpp \
            private/geodatariversurveycrosssectionwindow_datatabledelegate.cpp \
            private/geodatariversurveycrosssectionwindow_riversurveytabledelegate.cpp \
